refactor(applescript): extract AppleScriptRefBuilder shared resolver (#104 Phase 0)
Phase 0 of the #104 disambiguation sweep. Extracts the account-reference
primitives into a shared module so PR-A..E become pure call-site swaps.

- NEW AppleScriptRefBuilder.swift:
  * mailboxRefByAccountId / msgRefByAccountId — moved verbatim from
    SaveAttachmentScriptBuilder.swift (same module, no import change)
  * resolveMailboxRef / resolveMsgRef — NEW: the single chokepoint.
    accountId non-nil/non-empty → (account id "<UUID>") UUID selector;
    nil/empty → legacy account "<display_name>" form, byte-identical
    to MailController.msgRef / mailboxRef pre-sweep output.
- SaveAttachmentScriptBuilder.swift: buildSaveAttachmentScript now
  delegates to resolveMsgRef (was an inlined if/else). Trimmed the
  file header to point at AppleScriptRefBuilder for the primitives.

#110 (escape-function dedup) intentionally NOT folded in — the new
builder uses the existing appleScriptEscape free function; it does not
touch MailController.escapeForAppleScript. Kept separate per plan
Decision Point 1.

swift test: 16/16 (AppleScriptRefBuilderTests 9 + SaveAttachmentScriptBuilderTests 7).
Refs #104
